RENAME: Change PromptProvider to prompt_provider
Renamed PromptProvider module to follow Python's PEP 8 naming conventions. Updated all import statements in run_agent_client.py accordingly.
